refactor(mrv2): HookedGPUModelRunner fork, minimal DllmGPUModelRunner
- Add vllm_gpu_model_runner_fork.HookedGPUModelRunner tracking v0.20.0
  prepare_inputs and sample_tokens with hooks (expand map, PP width, speculator
  phase, before_execute_model) and execute_model kwargs forwarding.
- Shrink DllmGPUModelRunner to sample(), take_dllm_draft_token_ids(), and hook
  overrides; keep _GPUModelRunnerPrepareInputsFork as alias.
- Update DESIGN_MVP and test_two_phase_dllm assertions.

Made-with: Cursor
